test(navigation): split monolithic test files into per-component modules
Both `tests/proptest_state_machines/navigation.rs` (1,062 lines) and
`tests/spec_conformance/navigation.rs` (266 lines) held every navigation
component's tests in a single file. Splitting them mirrors the proven
`input/` layout (PR #692): shared strategies live in the directory's
`mod.rs`, each component owns a `<component>.rs` sibling, and the root
test binary repoints at `navigation/mod.rs`.

- Proptests: shared `arb_key`/`arb_orientation`/`arb_direction` in
  `navigation/mod.rs`; per-component `accordion.rs`, `pagination.rs`,
  `steps.rs`, `tabs.rs`, `tree_view.rs` (verbatim move of their
  blocks). `breadcrumbs` is stateless (no state-machine proptest).
- Spec-conformance: all 8 components get a sibling module under
  `navigation/`, lifted verbatim from the monolith.

Refs #223

Entire-Checkpoint: 57e53ced2718
