+++ crates/ars-components/tests/spec_conformance/navigation/tree_view.rs
@@ -1,0 +1,59 @@
+use ars_collections::Key;
+use ars_components::navigation::tree_view;
+
+use crate::helper::assert_anatomy;
+
+#[test]
+fn tree_view_anatomy_matches_spec() {
+    // Spec references:
+    // - `spec/components/navigation/tree-view.md` §2 base anatomy declares
+    //   Root / Branch / BranchControl / BranchIndicator / BranchText /
+    //   BranchContent / Leaf / LeafText.
+    // - §4 (drag-and-drop reorder variant) layers the DragHandle and
+    //   DropIndicator parts on top, appended last.
+    assert_anatomy(
+        "tree-view",
+        &[
+            (tree_view::Part::Root, "root"),
+            (
+                tree_view::Part::Branch {
+                    node_id: Key::default(),
+                },
+                "branch",
+            ),
+            (
+                tree_view::Part::BranchControl {
+                    node_id: Key::default(),
+                },
+                "branch-control",
+            ),
+            (
+                tree_view::Part::BranchIndicator {
+                    node_id: Key::default(),
+                },
+                "branch-indicator",
+            ),
+            (tree_view::Part::BranchText, "branch-text"),
+            (
+                tree_view::Part::BranchContent {
+                    node_id: Key::default(),
+                },
+                "branch-content",
+            ),
+            (
+                tree_view::Part::Leaf {
+                    node_id: Key::default(),
+                },
+                "leaf",
+            ),
+            (tree_view::Part::LeafText, "leaf-text"),
+            (
+                tree_view::Part::DragHandle {
+                    node_id: Key::default(),
+                },
+                "drag-handle",
+            ),
+            (tree_view::Part::DropIndicator, "drop-indicator"),
+        ],
+    );
+}
